Recursively enumerate directory trees--sort of
Recursive enumeration is working, but the process doesn't properly exit
when done, as the work channel isn't closed properly.

I need to figure out a way to signal when all the enumerations are done
so that the worker thread will drop is thread_work_sender and close the
work channel.

It probably will be easiest to have the main thread enumerate the
directory and only enqueue File items as work. This may also simplify
error handling.