Add clone of Sender<Work> to Work::Directory
Currently, the process hangs after directory enumeration is done. This
is because thread's `thead_work_sender: Sender<Work>` keeps the sending
side of the channel open.

Introduce a Work::Directory-local clone of the Sender<Work> that the
Work can use to enqueue follow up work if needed. Remove
`thread_work_sender`. Now, when all directory enumeration work has
completed, all the copies of `Sender<Work>` will be dropped, and the
sending side of the work channel will be closed.